requirements.txt
asgiref==3.3.4
astroid==2.4.2
autopep8==1.5.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pycodestyle==2.6.0
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
